requirements.txt
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
filelock==3.12.0
greenlet==2.0.2
idna==3.4
oscrypto==1.3.0
packaging==23.1
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.7.0
pyOpenSSL==23.1.1
pytz==2023.3
requests==2.30.0
snowflake-connector-python==3.0.3
snowflake-sqlalchemy==1.4.7
SQLAlchemy==1.4.48
typing_extensions==4.5.0
urllib3==1.26.15
